# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SMM API
  version: 0.1.0
  description: |
    Social Media Management API. This spec is the single source of truth for the
    HTTP contract: Go server/handler types are generated with `oapi-codegen` and
    FE types with `openapi-typescript` into `packages/shared`. Do not hand-write
    DTOs on either side.
servers:
  - url: http://localhost:24080
    description: Local development
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token and mint a new access token
      operationId: refresh
      responses:
        '200':
          description: New session. Rotates the active refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    post:
      tags: [accounts]
      summary: Pause or resume an account
      description: |
        `pause` suspends use (the session survives on the container PVC);
        `resume` reactivates without a fresh login.
      operationId: setAccountStatus
      security:
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAccountStatusRequest'
      responses:
        '200':
          description: Account updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [accounts]
      summary: Remove an account
      description: |
        Releases the container slot (an AUTO container that empties is reaped)
        and deletes the account row.
      operationId: deleteAccount
      security:
        - accessCookie: []
      responses:
        '204':
          description: Account removed.
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/audit:
    get:
      tags: [audit]
      summary: List the audit trail
      description: |
        Returns who did what, when: one row per state-changing call
        (account/container/template/user create/update/delete, action
        enqueue), newest first, with the actor's email resolved and the
        request IP. Filters are all optional. Requires the `read` permission.
      operationId: listAudit
      security:
        - accessCookie: []
      parameters:
        - name: actorId
          in: query
          schema:
            type: string
          description: Filter to one user's actions.
        - name: action
          in: query
          schema:
            type: string
          description: e.g. `account.create`.
        - name: entity
          in: query
          schema:
            type: string
          description: e.g. `account`, `worker`, `template`.
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound on the entry time.
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Exclusive upper bound on the entry time.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
          description: Page size.
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page offset.
      responses:
        '200':
          description: Audit entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /api/users:
    get:
      tags: [users]
      summary: List team members
      description: |
        Returns every user in the single-team MVP with their role. Requires
        the `read` permission.
      operationId: listUsers
      security:
        - accessCookie: []
      responses:
        '200':
          description: Users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [users]
      summary: Create / invite a team member
      description: |
        Adds a user with a provisional password the operator shares out of
        band (the MVP has no email transport wired for invites). The password
        is argon2id-hashed; it is never returned. Requires `admin` (OWNER).
      operationId: createUser
      security:
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [users]
      summary: Update a team member
      description: |
        Changes a member's role and/or name. Changing a role revokes that
        user's refresh sessions so the new role applies immediately. Requires
        `admin` (OWNER).
      operationId: updateUser
      security:
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [users]
      summary: Remove a team member
      description: |
        Deletes the user. The last OWNER cannot be removed (the team must keep
        an admin), and an account cannot delete itself. Requires `admin`.
      operationId: deleteUser
      security:
        - accessCookie: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
      description: The account ID.
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: The user ID.
...
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: smm_at
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    LoginResponse:
      type: object
      required: [accessExpiry, user]
      properties:
        accessExpiry:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
    Role:
      type: string
      enum: [OWNER, STRATEGIST, OPERATOR, ANALYST]
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    Account:
      type: object
      required: [id, platform, username, authStatus, status]
      properties:
        id:
          type: string
        platform:
          type: string
          enum: [instagram, threads, facebook, linkedin, x, youtube, tiktok]
        username:
          type: string
        handle:
          type: string
          nullable: true
        authStatus:
          type: string
          enum: [AUTHENTICATING, NEEDS_INPUT, AUTHENTICATED, FAILED, QUARANTINED]
        status:
          type: string
          enum: [ACTIVE, PAUSED, ARCHIVED, DEAD]
        workerId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        lastError:
          type: string
          nullable: true
...
    CreateUserRequest:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        password:
          type: string
          format: password
          description: Write-only; argon2id-hashed, never returned.
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
    UserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    AuditLog:
      type: object
      description: One immutable audit-trail entry.
      required: [id, action, entity, entityId, result, ts]
      properties:
        id:
          type: string
        actorId:
          type: string
          description: Empty for system-initiated rows.
        actor:
          type: string
          description: The user's email, or `system`.
        action:
          type: string
          description: e.g. `account.create`.
        entity:
          type: string
          description: e.g. `account`, `worker`, `template`.
        entityId:
          type: string
        result:
          type: string
          description: '`ok`, or the failure reason.'
        ip:
          type: string
          description: The request IP; empty when unknown.
        ts:
          type: string
          format: date-time
    AuditLogList:
      type: object
      required: [rows, total, limit, offset]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        total:
          type: integer
          description: Entries matching the same filter, for pagination.
        limit:
          type: integer
        offset:
          type: integer
    SetAccountStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACTIVE, PAUSED]
...